aws-lambda-linkedIn-scraper/lambda/requirements.txt
# --- Core Libraries ---

#boto3 is required to interact with S3, DynamoDB, and SNS
boto3==1.34.83

#requests and beautifulsoup4 are for web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Optional: Add tabulate or pandas for formatting/debugging
tabulate==0.9.0
pandas==2.2.2

boto3
python-dotenv

#pip install -r requirements.txt
 



#Step-by-Step: Proper Installation of AWS CLI
# Step 1: Download AWS CLI v2 installer
#curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"

# Step 2: Unzip the installer
#unzip awscliv2.zip

# Step 3: Run the installer
#sudo ./aws/install